ui!: Provide reports on the Charts page as per role
[#OCD-4940]
