+++ src/app/api/reports.jsx
@@ -2,27 +2,14 @@ import { useQuery } from 'react-query';
 
 import { useAxios } from './axios';
 
-const useFetchReportGroupMetadata = (reportGroup) => {
+const useFetchReportMetadata = () => {
   const axios = useAxios();
-  return useQuery(['report-metadata', reportGroup], async () => {
-    const response = await axios.get(`report-data/report-metadata/group/${reportGroup}`);
+  return useQuery(['report-metadata'], async () => {
+    const response = await axios.get('report-metadata');
     return response.data;
-  }, {
-    enabled: !!reportGroup,
-  });
-};
-
-const useFetchReportMetadata = (report) => {
-  const axios = useAxios();
-  return useQuery(['report-metadata', report], async () => {
-    const response = await axios.get(`report-data/report-metadata/${report}`);
-    return response.data;
-  }, {
-    enabled: !!report,
   });
 };
 
 export {
-  useFetchReportGroupMetadata,
   useFetchReportMetadata,
 };
