+++ src/app/navigation/navigation-top.html
@@ -91,7 +91,6 @@
               <li ng-if="vm.hasAnyRole(['chpl-admin', 'chpl-onc', 'chpl-cms-staff'])"><a ui-sref="administration.cms"><i class="fa fa-caret-right" ng-if="vm.isActive('administration.cms')"></i> CMS</a></li>
               <li ng-if="vm.hasAnyRole(['chpl-admin', 'chpl-onc', 'chpl-onc-acb'])"><a ui-sref="administration.change-requests"><i class="fa fa-caret-right" ng-if="vm.isActive('administration.change-requests')"></i> Change Requests</a></li>
               <li ng-if="vm.hasAnyRole(['chpl-admin', 'chpl-onc', 'chpl-onc-acb'])"><a ui-sref="administration.system-maintenance"><i class="fa fa-caret-right" ng-if="vm.isActive('administration.system-maintenance')"></i> System Maintenance</a></li>
-              <li ng-if="vm.hasAnyRole(['chpl-admin', 'chpl-onc'])"><a ui-sref="dashboard"><i class="fa fa-caret-right" ng-if="vm.isActive('sys-admin.dashboard')"></i>Dashboard</a></li>
               <li ng-if="vm.hasAnyRole(['chpl-admin', 'chpl-onc', 'chpl-onc-acb'])"><a ui-sref="administration.url-checker"><i class="fa fa-caret-right" ng-if="vm.isActive('administration.url-checker')"></i>URL Checker</a></li>
               <li ng-if="vm.hasAnyRole(['chpl-admin'])"><a href="rest/ff4j-console/home">FF4j</a></li>
             </ul>
