+++ src/app/pages/resources/resources.module.js
@@ -1,7 +1,6 @@
 import { reactToAngularComponent } from '../../services/angular-react-helper';
 
 import ChplCmsLookupWrapper from './cms-lookup/cms-lookup-wrapper';
-import ChplDashboardWrapper from './dashboard/dashboard-wrapper';
 import ChplForgotPassword from './forgot-password/forgot-password-wrapper';
 import ChplNotFound from './not-found/not-found';
 import ChplResourcesApi from './api/api-wrapper';
@@ -20,5 +19,4 @@ angular
   .component('chplNotFoundBridge', reactToAngularComponent(ChplNotFound))
   .component('chplResourcesApiBridge', reactToAngularComponent(ChplResourcesApi))
   .component('chplResourcesDownloadWrapperBridge', reactToAngularComponent(ChplResourcesDownloadWrapper))
-  .component('chplResourcesOverviewBridge', reactToAngularComponent(ChplResourcesOverview))
-  .component('chplDashboardWrapperBridge', reactToAngularComponent(ChplDashboardWrapper));
+  .component('chplResourcesOverviewBridge', reactToAngularComponent(ChplResourcesOverview));
